# s1_reader.py: implement load_bursts speedups