--- a/.circleci/freeze-deps.sh
+++ b/.circleci/freeze-deps.sh
@@ -2,5 +2,5 @@
 docker run -v "$PWD:/mnt" -w /mnt --rm -it continuumio/miniconda3 bash -c '\
     conda config --add channels conda-forge && \
     conda config --remove channels defaults && \
-    conda update conda && \
-    conda install -y -c conda-forge isce3=0.8 build setuptools pytest isce3 shapely && conda list --explicit > specfile.txt'
+    conda update -y conda && \
+    conda install -y -c conda-forge isce3=0.8 build setuptools pytest isce3 shapely lxml && conda list --explicit > specfile.txt'