--- a/.gitignore
+++ b/.gitignore
@@ -116,6 +116,9 @@ venv.bak/
 .spyderproject
 .spyproject
 
+# VScode project settings
+.vscode
+
 # Rope project settings
 .ropeproject