--- a/requirements.txt
+++ b/requirements.txt
@@ -5,3 +5,4 @@ packaging
 requests
 #isce3    # since the conda-installed isce3 is not the most updated version, installing isce3 from stratch is recommended, to stay in sync with isce3 development.
 shapely
+lxml