… .circleci/specfile.txt