I added three speedups which make loading 1 subswath run about 3x faster:

1. use fromisoformat instead of strptime (https://stackoverflow.com/a/61710371/4174466)

2. use lxml to parse if available (https://stackoverflow.com/a/18507562/4174466)

3. use eval poly1d by passing array instead of a comprehension

I was checking on this with the following little program to load one subswath's bursts:
```python
(mapping) [staniewi@aurora hawaii]$ cat tests1.py
import s1reader
import time

tstart = time.time()
b = s1reader.load_bursts("data/S1B_IW_SLC__1SDV_20200927T161542_20200927T161609_023563_02CC58_C487.zip", "orbits/S1B_OPER_AUX_POEORB_OPOD_20210317T172346_V20200926T225942_20200928T005942.EOF", 2)
```

First, with the current code:
```
$ python tests1.py
Elapsed: 775.102 ms
```

With change (1), the fromisoformat

```
$ python tests1.py
Elapsed: 577.467 ms
```

With change (2), the lxml parser
```
$ python tests1.py
Elapsed: 319.141 ms
```

With (3), with the Poly1D evaluation change using isce3 instead of the list comprehension:

```
$ python tests1.py
Elapsed: 257.990 ms
```