@scottstanie 2022-10-17
Just a note on using lxml optionally- although it is quite common, it does involve a separate install. So the question is whether we think the 1.5x - 2x speedup is worth adding that in.