@LiangJYu 2022-10-17
> Just a note on using lxml optionally- although it is quite common, it does involve a separate install. So the question is whether we think the 1.5x - 2x speedup is worth adding that in.

I'm ok with having `lxml` as optional. It's a nice speed if anyone who only wants to read a bunch of SAFEs for whatever reason. I'm guessing in the context of running a workflow the improvement in processing it is not as significant?