@scottstanie 2022-10-17
> I'm guessing in the context of running a workflow the improvement in processing is not as significant?

Yes I believe that it's small for a workflow of one burst in COMPASS. The this PR came about was for the work i'm doing on the `s1_geo_stack.py`  script here: https://github.com/opera-adt/COMPASS/pull/53 . For a large stack, it took about 10 minutes to run, and 99% of the time was in `load_bursts`.