@yunjunz 2022-10-17
`lxml` is light and very commonly used. If it speeds up our code, it's perfectly fine with me to add it to the `requirements.txt` file.