@scottstanie 2022-10-18
Alright I took out the `try/except` in favor of adding it to the requirements file.