@yunjunz 2022-10-27
> Do we want to follow that and swap lxml for [defusedxml](https://github.com/tiran/defusedxml)? Or do we think that it's alright because the only XML's we're parsing are from Sentinel-1, rather than accepting arbitrary XML?

Yeah, I think we are parsing S1 XML only, I think we are fine using `lxml`, without `defusedxml`.